Add support for Swift Package Manager
