+++ Family.podspec
@@ -1,7 +1,7 @@
 Pod::Spec.new do |s|
   s.name             = "Family"
   s.summary          = "A child view controller framework that makes setting up your parent controllers as easy as pie."
-  s.version          = "2.1.2"
+  s.version          = "2.2.0"
   s.homepage         = "https://github.com/zenangst/Family"
   s.license          = 'MIT'
   s.author           = { "Christoffer Winterkvist" => "christoffer@winterkvist.com" }
@@ -17,9 +17,9 @@ Pod::Spec.new do |s|
   s.tvos.deployment_target = '10.0'
 
   s.requires_arc = true
-  s.ios.source_files = 'Sources/{iOS+tvOS,Shared}/**/*'
-  s.tvos.source_files = 'Sources/{iOS+tvOS,Shared}/**/*'
-  s.macos.source_files = 'Sources/{macOS,Shared}/**/*'
+  s.ios.source_files = 'Sources/{UIKit,Shared}/**/*'
+  s.tvos.source_files = 'Sources/{UIKit,Shared}/**/*'
+  s.macos.source_files = 'Sources/{AppKit,Shared}/**/*'
 
   s.pod_target_xcconfig = { 'SWIFT_VERSION' => '5.0' }
 end
