+++ Sources/UIKit/CALayer+Extensions.swift
@@ -1,3 +1,4 @@
+#if canImport(UIKit)
 import UIKit
 
 extension CALayer {
@@ -31,3 +32,4 @@ extension CALayer {
     return nil
   }
 }
+#endif
